common: move pxEvents.h to main project
